+++ notify/src/events.rs
@@ -1,5 +1,8 @@
 use super::scope::Scope;
+use crate::error::Error;
+use serde::{Deserialize, Serialize};
 use std::ops::{Index, IndexMut};
+use std::str::FromStr;
 use workflow_core::enums::usize_try_from;
 
 macro_rules! event_type_enum {
@@ -25,21 +28,41 @@ macro_rules! event_type_enum {
 }
 
 event_type_enum! {
-#[derive(Clone, Copy, Debug, Eq, PartialEq, Hash)]
-pub enum EventType {
-    BlockAdded = 0,
-    VirtualChainChanged,
-    FinalityConflict,
-    FinalityConflictResolved,
-    UtxosChanged,
-    SinkBlueScoreChanged,
-    VirtualDaaScoreChanged,
-    PruningPointUtxoSetOverride,
-    NewBlockTemplate,
-}
+    #[derive(Clone, Copy, Debug, Eq, PartialEq, Hash, Serialize, Deserialize)]
+    #[serde(rename_all = "kebab-case")]
+    pub enum EventType {
+        BlockAdded = 0,
+        VirtualChainChanged,
+        FinalityConflict,
+        FinalityConflictResolved,
+        UtxosChanged,
+        SinkBlueScoreChanged,
+        VirtualDaaScoreChanged,
+        PruningPointUtxoSetOverride,
+        NewBlockTemplate,
+    }
 }
 
 pub const EVENT_COUNT: usize = 9;
+
+impl FromStr for EventType {
+    type Err = Error;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s {
+            "block-added" => Ok(EventType::BlockAdded),
+            "virtual-chain-changed" => Ok(EventType::VirtualChainChanged),
+            "finality-conflict" => Ok(EventType::FinalityConflict),
+            "finality-conflict-resolved" => Ok(EventType::FinalityConflictResolved),
+            "utxos-changed" => Ok(EventType::UtxosChanged),
+            "sink-blue-score-changed" => Ok(EventType::SinkBlueScoreChanged),
+            "virtual-daa-score-changed" => Ok(EventType::VirtualDaaScoreChanged),
+            "pruning-point-utxo-set-override" => Ok(EventType::PruningPointUtxoSetOverride),
+            "new-block-template" => Ok(EventType::NewBlockTemplate),
+            _ => Err(Error::InvalidEventType(s.to_string())),
+        }
+    }
+}
 
 /// Generic array with [`EventType`] strongly-typed index
 #[derive(Default, Clone, Copy, Debug)]