+++ rpc/wrpc/wasm/src/imports.rs
@@ -1,5 +1,6 @@
 #![allow(unused_imports)]
 
+pub use ahash::AHashMap;
 pub use cfg_if::cfg_if;
 pub use futures::*;
 pub use js_sys::Function;
@@ -23,6 +24,7 @@ pub use kaspa_wrpc_client::client::*;
 pub use kaspa_wrpc_client::error::Error;
 pub use kaspa_wrpc_client::result::Result;
 pub use serde::{Deserialize, Serialize};
+pub use std::str::FromStr;
 pub use std::sync::{
     atomic::{AtomicBool, Ordering},
     Arc, Mutex,