+++ wallet/bip32/src/mnemonic/language.rs
@@ -9,9 +9,13 @@ use super::bits::{Bits, Bits11};
 use std::{collections::BTreeMap, vec::Vec};
 use wasm_bindgen::prelude::*;
 
-/// BIP39 Supported languages.
+/// 
+/// Languages supported by BIP39.
 ///
-/// Presently only English is specified by the BIP39 standard
+/// Presently only English is specified by the BIP39 standard.
+///
+/// @see {@link Mnemonic}
+///
 /// @category Wallet SDK
 #[derive(Copy, Clone, Debug, Default)]
 #[wasm_bindgen]