+++ wasm/src/lib.rs
@@ -40,8 +40,8 @@ available NPM modules:
 The `kaspa-wasm` module is a pure WASM32 module that includes
 the entire wallet framework, but does not support RPC due to an absence
 of a native WebSocket in NodeJs environment, while
-the `kaspa` module includes `isomorphic-ws` dependency simulating
-the W3C WebSocket and thus supports RPC.
+the `kaspa` module includes `websocket` package dependency simulating
+the W3C WebSocket and due to this supports RPC.
 
 ## Examples
 
@@ -56,20 +56,18 @@ repository at <https://github.com/kaspanet/rusty-kaspa/releases>.
 
 ## Using RPC
 
-**NODEJS:** If you are building from source, to use WASM RPC client in the NodeJS environment,
-you need to introduce a W3C WebSocket object before loading the WASM32 library. You can use
-any Node.js module that exposes a W3C-compatible WebSocket implementation. Two of such modules
-are [WebSocket](https://www.npmjs.com/package/websocket) (provides a custom implementation)
-and [isomorphic-ws](https://www.npmjs.com/package/isomorphic-ws) (built on top of the ws
-WebSocket module).
+**NODEJS:** If you are building from source, to use WASM RPC client
+in the NodeJS environment, you need to introduce a global W3C WebSocket
+object before loading the WASM32 library (to simulate the browser behavior).
+You can the [WebSocket](https://www.npmjs.com/package/websocket)
+module that offers W3C WebSocket compatibility and is compatible
+with Kaspa RPC implementation.
 
 You can use the following shims:
 
 ```js
 // WebSocket
 globalThis.WebSocket = require('websocket').w3cwebsocket;
-// isomorphic-ws
-globalThis.WebSocket = require('isomorphic-ws');
 ```
 
 ## Loading in a Web App
@@ -131,7 +129,10 @@ For more details, please follow the [**integrating with Kaspa**](https://kaspa.a
 
 #![allow(unused_imports)]
 
-#[cfg(all(any(feature = "wasm32-sdk", feature = "wasm32-rpc", feature = "wasm32-core", feature = "wasm32-keygen"), not(target_arch = "wasm32")))]
+#[cfg(all(
+    any(feature = "wasm32-sdk", feature = "wasm32-rpc", feature = "wasm32-core", feature = "wasm32-keygen"),
+    not(target_arch = "wasm32")
+))]
 compile_error!("`kaspa-wasm` crate for WASM32 target must be built with `--features wasm32-sdk|wasm32-rpc|wasm32-core|wasm32-keygen`");
 
 mod version;